Fixed bug where ibv read would end early
This isn't the nicest way of solving the problem, but for now it
works. The problem was that devices tend to write things in 1KB
blocks, and then there is sometimes a delay. During this delay, there
would be no data, so the ibvread call would end without pushing data
to the uio struct. So we really need to wait for a bit to make sure
there isn't anymore. I used tsleep for this.