IBV11 IEEE-488 controller driver  (#1)
* Porting IBV11 from RetroBSD/2.11BSD to this repo

I didn't like that RetroBSD contained a bunch of object files. It also
didn't have things like /etc so didn't feel like a full
distribution. This fork seems like a better place to be working from,
unfortunately lost the history though. Oh well - it's not like the
driver works yet anyway.

* Forgot to define these additional interrupt handlers

* Function headers

* character device definition in conf.c

* Removal of my own buffer stuff in favour of uio

* Roughly implemented bus comms

* interrupt handlers have to be defined

* Write interrupts are now working.

Still have issues with read interrupts. I'm not sure if the message
I'm sending out is invalid and not worth of a result, or if the read
interrupts are being ignored for some reason.  Are they even lower
priority? I don't think so...

* IBV11 driver 'works'

There are still some bugs to fix which I will detail in the pull
request.

* NIBV config option

* Fixed bug where ibv read would end early

This isn't the nicest way of solving the problem, but for now it
works. The problem was that devices tend to write things in 1KB
blocks, and then there is sometimes a delay. During this delay, there
would be no data, so the ibvread call would end without pushing data
to the uio struct. So we really need to wait for a bit to make sure
there isn't anymore. I used tsleep for this.